Fix bad recursion in multikey_multipivot_brute_simd{2,4}
For the last bucket the is_end() check is not required in the byte
alphabet case, because the last pivot may never be the NULL byte.

When dealing with superalphabets, even the last pivot may return true
for is_end(), ie. these strings end here, and we must not recurse into
those subinputs.